+++ packages/relay-signer/src/action/index.ts
@@ -54,7 +54,7 @@ export class ActionRelayer extends BaseActionClient implements IRelayer {
 
   public async replaceTransactionById(id: string, payload: RelayerTransactionPayload): Promise<RelayerTransaction> {
     const txPayload: RelayerTransactionPayload & { id: string } = { ...payload, id };
-    return this.execute({ action: 'replace-tx', txPayload });
+    return this.execute({ action: 'replace-tx', payload: txPayload });
   }
 
   public async replaceTransactionByNonce(
@@ -62,11 +62,12 @@ export class ActionRelayer extends BaseActionClient implements IRelayer {
     payload: RelayerTransactionPayload,
   ): Promise<RelayerTransaction> {
     const txPayload: RelayerTransactionPayload & { nonce: number } = { ...payload, nonce };
-    return this.execute({ action: 'replace-tx', payload });
+    return this.execute({ action: 'replace-tx', payload: txPayload });
   }
 
   public async cancelTransactionById(id: string): Promise<TransactionDeleteResponse> {
-    return this.execute({ action: 'cancel-tx', id });
+    const payload = id;
+    return this.execute({ action: 'cancel-tx', payload });
   }
 
   public async getRelayer(): Promise<RelayerGetResponse> {
@@ -85,6 +86,13 @@ export class ActionRelayer extends BaseActionClient implements IRelayer {
     return this.execute({
       action: 'get-tx' as const,
       payload: id,
+    });
+  }
+
+  public async getTransactionByNonce(nonce: number): Promise<RelayerTransaction> {
+    return this.execute({
+      action: 'get-tx' as const,
+      payload: String(nonce),
     });
   }