+++ examples/relayer-load-balance/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @openzeppelin/defender-sdk-example-relayer-load-balance
 
+## 1.15.1
+
+### Patch Changes
+
+- c68c41a: feat: Add support for Actions high frequency intervals
+  feat: Add support to approval process list
+  feat:Add address book client methods
+  fix: Upgrade web3 dependency to v4 in defender sdk
+  feat: Predefine auth config type for relayer signer client
+  feat: Add support to query transactions by nonce
+  fix: Type inference in relay signer
+  fix: Fix payload format for relay-signer
+- Updated dependencies [c68c41a]
+  - @openzeppelin/defender-sdk@1.15.1
+
 ## 1.15.0
 
 ### Minor Changes